+++ src/manifests/original-source.ts
@@ -1,0 +1,16 @@
+import { ManifestSourceType } from "../types/manifest";
+import { InputSource } from "./input-source";
+
+export class OriginalSource
+{
+    public kind: ManifestSourceType;
+    public path: string;
+
+    public innerSources : InputSource[] = [];
+    
+    constructor(kind: ManifestSourceType, path: string)
+    {
+        this.kind = kind;
+        this.path = path;
+    }
+}
